simplify headers
