+++ oneflow/core/control/ctrl_client.h
@@ -15,5 +15,5 @@ limitations under the License.
 */
 #ifndef ONEFLOW_CORE_CONTROL_CTRL_CLIENT_
 #define ONEFLOW_CORE_CONTROL_CTRL_CLIENT_
-#include "oneflow/core/rpc/include/local/ctrl.h"
+#include "oneflow/core/rpc/include/ctrl.h"
 #endif  // ONEFLOW_CORE_CONTROL_CTRL_CLIENT_
