--- oneflow/core/rpc/include/base/bootstrap.h
@@ -1,4 +1,0 @@
-#ifndef ONEFLOW_CORE_RPC_INCLUDE_BASE_BOOTSTRAP_H_
-#define ONEFLOW_CORE_RPC_INCLUDE_BASE_BOOTSTRAP_H_
-
-#endif  // ONEFLOW_CORE_RPC_INCLUDE_BASE_BOOTSTRAP_H_
