--- oneflow/core/rpc/include/base/rpc.h
@@ -1,32 +1,0 @@
-/*
-Copyright 2020 The OneFlow Authors. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#ifndef ONEFLOW_CORE_RPC_INCLUDE_BASE_RPC_
-#define ONEFLOW_CORE_RPC_INCLUDE_BASE_RPC_
-
-#include "oneflow/core/actor/actor_message.h"
-#include "oneflow/core/common/protobuf.h"
-#include "oneflow/core/control/control.pb.h"
-#include "oneflow/core/job/global_for.h"
-
-namespace oneflow {
-
-class RpcClientBase {};
-
-class RpcServerBase {};
-
-}  // namespace oneflow
-
-#endif  // ONEFLOW_CORE_RPC_INCLUDE_BASE_RPC_
