+++ oneflow/core/rpc/include/grpc.h
@@ -1,0 +1,4 @@
+#ifndef ONEFLOW_CORE_RPC_INCLUDE_GRPC_H_
+#define ONEFLOW_CORE_RPC_INCLUDE_GRPC_H_
+
+#endif  // ONEFLOW_CORE_RPC_INCLUDE_GRPC_H_
