+++ oneflow/core/rpc/include/local.h
@@ -13,16 +13,71 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_RPC_INCLUDE_LOCAL_CTRL_
-#define ONEFLOW_CORE_RPC_INCLUDE_LOCAL_CTRL_
+#ifndef ONEFLOW_CORE_RPC_INCLUDE_LOCAL_H_
+#define ONEFLOW_CORE_RPC_INCLUDE_LOCAL_H_
 
 #include <string>
 #include <unordered_map>
+#include "oneflow/core/common/protobuf.h"
 #include "oneflow/core/control/ctrl_bootstrap.pb.h"
-#include "oneflow/core/rpc/include/local/rpc.h"
-#include "oneflow/core/rpc/include/base/ctrl.h"
+#include "oneflow/core/rpc/include/base.h"
 
 namespace oneflow {
+
+class RpcClient : RpcClientBase {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(RpcClient);
+  virtual ~RpcClient() = default;
+
+  void Barrier(const std::string& barrier_name);
+  void Barrier(const std::string& barrier_name, int32_t barrier_num);
+
+  TryLockResult TryLock(const std::string& name);
+  void NotifyDone(const std::string& name);
+  void WaitUntilDone(const std::string& name);
+
+  void PushKV(const std::string& k, std::function<void(std::string*)> VSetter);
+  void PushKV(const std::string& k, const std::string& v);
+  void PushKV(const std::string& k, const PbMessage& msg);
+  void PushMasterKV(const std::string& k, const PbMessage& msg);
+  template<typename T>
+  typename std::enable_if<std::is_arithmetic<T>::value>::type PushKVT(const std::string& k, T v) {
+    PushKV(k, std::to_string(v));
+  }
+
+  void ClearKV(const std::string& k);
+  void ClearMasterKV(const std::string& k);
+
+  void PullKV(const std::string& k, std::function<void(const std::string&)> VGetter);
+  void PullKV(const std::string& k, std::string* v);
+  void PullKV(const std::string& k, PbMessage* msg);
+  void PullMasterKV(const std::string& k, PbMessage* msg);
+  template<typename T>
+  typename std::enable_if<std::is_arithmetic<T>::value>::type PullKVT(const std::string& k, T* v) {
+    std::string v_str;
+    PullKV(k, &v_str);
+    *v = oneflow_cast<T>(v_str);
+  }
+
+  void PushActEvent(const ActEvent&){};
+  void Clear();
+
+  int32_t IncreaseCount(const std::string& k, int32_t v);
+  int32_t IncreaseCount(const std::string& k) { return IncreaseCount(k, 1); }
+  void EraseCount(const std::string& k);
+
+ protected:
+  RpcClient(){};
+  void PushMasterKV(const std::string& k, std::function<void(std::string*)> VSetter);
+  void PullMasterKV(const std::string& k, std::function<void(const std::string&)> VGetter);
+
+  HashSet<std::string> done_names_;
+  std::mutex done_names_mtx_;
+  std::condition_variable done_names_cv_;
+  HashMap<std::string, std::string> kv_;
+  std::mutex kv_mtx_;
+  std::condition_variable kv_cv_;
+};
 
 class CtrlClient final : public RpcClient {
  public:
@@ -37,6 +92,14 @@ class CtrlClient final : public RpcClient {
   const ProcessCtx& process_ctx() const { return process_ctx_; }
 
   ProcessCtx process_ctx_;
+};
+
+class LocalRpcManager {
+ public:
+  LocalRpcManager() {}
+  virtual ~LocalRpcManager() {}
+  virtual void CreateClient() {}
+  virtual void TearDown() {}
 };
 
 #define FILE_LINE_STR __FILE__ ":" OF_PP_STRINGIZE(__LINE__)
@@ -118,4 +181,4 @@ class CtrlCall final : public CtrlCallIf {
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_RPC_INCLUDE_LOCAL_CTRL_
+#endif  // ONEFLOW_CORE_RPC_INCLUDE_LOCAL_H_
